Switch from lattice.dispatch to ct.dispatch (#55)
* Switch from lattice.dispatch to ct.dispatch

* reorganized dispatcher a bit

* Created files for dispatcher plugins

* Implemented dispatcher plugins and removed extra class from executor

* Added dispatch functions to covalent's init

* Removed dispatch functions from lattice, tests will fail

* Made dispatch and dispatch_sync functions static

* Stopped original lattice manipulation when dispatching with a copy instead

* Updating tests

* Updated more tests, sublatticing is failing

* Fixed sublattice test and dispatch_sync

* Updated tests to use ct.dispatch and delete results once test is done

* Updated howtos for ct.dispatch

* Updated tutorials with ct.dispatch

* Updated README and rst files

* Added new interface to api.rst

* Added docstrings and removed dispatcher constraint from lattice

* Updated the api.rst and docstring for dispatch and dispatch_sync

* Updated a test due to reflect a refactor of dispatcher/core/init

* Fixed tests failing

* Added tests for newly created dispatcher classes/functions

* Updated changelog and version

* Updated concepts.rst

* Deprecation warning being shown in lattice dispatching methods and updated changelog
