+++ doc/source/getting_started/index.rst
@@ -156,7 +156,7 @@ Let's look at a simple example to get started with Covalent. Before starting, en
        return excitement(phrase)
 
    # Dispatch the workflow
-   dispatch_id = simple_workflow.dispatch("Hello", "World")
+   dispatch_id = ct.dispatch(simple_workflow)("Hello", "World")
 
 Navigate to the Covalent UI at `<http://0.0.0.0:47007>`_ to see your workflow in the queue:
