+++ tests/covalent_dispatcher_tests/_core/init_test.py
@@ -23,11 +23,10 @@ Tests for the core functionality of the dispatcher.
 """
 
 import cloudpickle as pickle
-import covalent as ct
 
-from covalent_dispatcher._core import _plan_workflow
+import covalent as ct
 from covalent._results_manager import Result
-from covalent._results_manager import results_manager as rm
+from covalent_dispatcher._core.execution import _plan_workflow
 
 TEST_RESULTS_DIR = "/tmp/results"
 
@@ -43,7 +42,6 @@ def get_mock_result() -> Result:
         print("Error!", file=sys.stderr)
         return x
 
-
     @ct.lattice(results_dir=TEST_RESULTS_DIR)
     def pipeline(x):
         return task(x)
@@ -58,8 +56,7 @@ def get_mock_result() -> Result:
 
 
 def test_plan_workflow():
-    """Test workflow planning method.
-    """
+    """Test workflow planning method."""
 
     mock_result = get_mock_result()
     _plan_workflow(result_object=mock_result)
@@ -67,4 +64,4 @@ def test_plan_workflow():
     # Updated transport graph post planning
     updated_tg = pickle.loads(mock_result.lattice.transport_graph.serialize(metadata_only=True))
 
-    assert 'results_dir' in updated_tg['nodes'][0]['exec_plan'].execution_args
+    assert "results_dir" in updated_tg["nodes"][0]["exec_plan"].execution_args
